requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
redis
celery
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-dotenv
stripe
pytest
pytest-asyncio
httpx